[DCB] refactor page objects into single file
