--- src/test/scala/uk/gov/hmrc/test/ui/pages/WhatIsNinoChequePage.scala
@@ -1,27 +1,0 @@
-/*
- * Copyright 2024 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package uk.gov.hmrc.test.ui.pages
-
-import uk.gov.hmrc.test.ui.conf.TestConfiguration
-
-object WhatIsNinoChequePage extends BasePage {
-
-  val h1    = "What is your National Insurance number?"
-  val url   = TestConfiguration.url("p800-refunds-frontend") + "/cheque/enter-your-national-insurance-number"
-  val title = "Cheque - enter your National Insurance number"
-
-}
